Create epa_air_data_daily_air_quality_ozone_pm25_austin_roundrock.csv
https://www.epa.gov/outdoor-air-quality-data/air-data-daily-air-quality-tracker
The most recent AQI values are based on available ozone and PM2.5 data from AirNow (see "Use of AirNow Data" info box).  The 20-year high/low backdrop is based on AQS data from 1980-2019.  It shows the range of the highest and lowest daily max AQI values on each date.  The 5-year median is based on AQS data from 2015-2019.  It  shows the median of daily max AQI values on each date.